2_Assignment/requirements.txt
datasets
transformers
scikit-learn
sentencepiece
torch
protobuf
accelerate
jsonlines
pillow